✨ feat(treapArray):

+++ 算法竞赛进阶指南/GoDS (Go Data Structures)/src/math/matrixquickpow/matrixquickpow_test.go
@@ -11,8 +11,8 @@ import (
 
 func TestMatrix_Pow(t *testing.T) {
 	type args struct {
-		exp int64
-		mod int64
+		exp int
+		mod int
 	}
 	tests := []struct {
 		name   string
